refactor(frontend): split app into components and composables
